chore: move shared chat modules out of routes
