+++ apps/web/src/features/chat/components/chat-web-search-toggle.tsx
@@ -8,6 +8,7 @@ import { stellaToast } from "@stll/ui/components/toast";
 import { cn } from "@stll/ui/lib/utils";
 
 import Tooltip from "@/components/tooltip";
+import { invalidateChatThread } from "@/features/chat/queries";
 import { api } from "@/lib/api";
 import type { ChatThreadRef } from "@/lib/chat-thread-ref";
 import { useChatWebSearchPreferenceStore } from "@/lib/chat-web-search-store";
@@ -15,7 +16,6 @@ import { detached } from "@/lib/detached";
 import { unwrapEden } from "@/lib/errors/api";
 import { userErrorFromThrown } from "@/lib/errors/user-safe";
 import { toSafeId } from "@/lib/safe-id";
-import { invalidateChatThread } from "@/routes/_protected.chat/-queries";
 
 type ChatWebSearchToggleProps = {
   enabled: boolean;
