+++ apps/web/src/features/chat/components/suggested-followup-chips.tsx
