+++ apps/web/src/features/chat/hooks/use-global-chat-mention-registration.ts
