+++ apps/web/src/features/chat/hooks/use-workspace-chat-mention-registration.ts
