+++ apps/web/src/features/chat/lib/build-chat-request-message.ts
