+++ scripts/ratchet-baseline.json
@@ -251,6 +251,8 @@
       "apps/web/src/features/case-law/components/case-viewer/analysis/types.ts": 2,
       "apps/web/src/features/case-law/components/case-viewer/decision-text.tsx": 1,
       "apps/web/src/features/case-law/components/case-viewer/decision-workspace.tsx": 1,
+      "apps/web/src/features/chat/hooks/use-chat-session.ts": 4,
+      "apps/web/src/features/chat/queries.ts": 1,
       "apps/web/src/hooks/external-file-drop.logic.ts": 5,
       "apps/web/src/hooks/use-effect.ts": 4,
       "apps/web/src/lib/analytics/posthog.ts": 1,
@@ -268,8 +270,6 @@
       "apps/web/src/lib/pdf/uint8array-to-hex-polyfill.logic.ts": 2,
       "apps/web/src/lib/slot.tsx": 1,
       "apps/web/src/router.tsx": 1,
-      "apps/web/src/routes/_protected.chat/-hooks/use-chat-session.ts": 4,
-      "apps/web/src/routes/_protected.chat/-queries.ts": 1,
       "apps/web/src/routes/_protected.chat/index.tsx": 1,
       "apps/web/src/routes/_protected.knowledge/-components/clause-body.tsx": 2,
       "apps/web/src/routes/_protected.knowledge/-components/clause-diff-view.tsx": 2,
@@ -440,14 +440,12 @@
     }
   },
   "cross-route-private-imports": {
-    "count": 217,
+    "count": 182,
     "files": {
       "apps/web/src/components/ai-prompt-input/ai-prompt-input.tsx": 2,
-      "apps/web/src/components/ai-suggestions/file-chat-overlay.tsx": 8,
-      "apps/web/src/components/ai-suggestions/review-panel.impl.tsx": 1,
-      "apps/web/src/components/ai-suggestions/use-review-actions.ts": 1,
-      "apps/web/src/components/app-sidebar.tsx": 9,
-      "apps/web/src/components/breadcrumbs/chat-breadcrumb.tsx": 2,
+      "apps/web/src/components/ai-suggestions/file-chat-overlay.tsx": 1,
+      "apps/web/src/components/app-sidebar.tsx": 8,
+      "apps/web/src/components/breadcrumbs/chat-breadcrumb.tsx": 1,
       "apps/web/src/components/breadcrumbs/clauses-breadcrumb.tsx": 1,
       "apps/web/src/components/breadcrumbs/contact-breadcrumb.tsx": 1,
       "apps/web/src/components/breadcrumbs/pdf-breadcrumb.tsx": 1,
@@ -461,12 +459,9 @@
       "apps/web/src/components/chat-mention-list.tsx": 1,
       "apps/web/src/components/chat-mention-node.tsx": 1,
       "apps/web/src/components/chat-mention-providers.tsx": 2,
-      "apps/web/src/components/chat/chat-approval-context.tsx": 1,
-      "apps/web/src/components/chat/chat-composer-dock.tsx": 2,
       "apps/web/src/components/chat/chat-draft-attachment-chips.tsx": 1,
       "apps/web/src/components/chat/chat-matter-picker.tsx": 1,
-      "apps/web/src/components/chat/chat-thread-messages.tsx": 1,
-      "apps/web/src/components/chat/composer-plus-menu.tsx": 5,
+      "apps/web/src/components/chat/composer-plus-menu.tsx": 4,
       "apps/web/src/components/chat/entity-link.tsx": 2,
       "apps/web/src/components/chat/entity-open.ts": 1,
       "apps/web/src/components/chat/source-chips.tsx": 3,
@@ -476,7 +471,7 @@
       "apps/web/src/components/chat/use-request-chat-about.ts": 1,
       "apps/web/src/components/drag-preview-content.tsx": 1,
       "apps/web/src/components/inspector/anonymization-facet.tsx": 2,
-      "apps/web/src/components/inspector/chat-tab-panel.tsx": 7,
+      "apps/web/src/components/inspector/chat-tab-panel.tsx": 1,
       "apps/web/src/components/inspector/document-ai-source-bar.tsx": 5,
       "apps/web/src/components/inspector/email-html-viewer.tsx": 1,
       "apps/web/src/components/inspector/entity-metadata-panel.tsx": 8,
@@ -486,8 +481,7 @@
       "apps/web/src/components/inspector/inspector-panel.tsx": 5,
       "apps/web/src/components/inspector/inspector-rail.tsx": 4,
       "apps/web/src/components/inspector/inspector-tab-header.tsx": 1,
-      "apps/web/src/components/inspector/maximize-tab.ts": 2,
-      "apps/web/src/components/inspector/playbook-facet.tsx": 3,
+      "apps/web/src/components/inspector/playbook-facet.tsx": 2,
       "apps/web/src/components/inspector/skill-resource-panel.tsx": 1,
       "apps/web/src/components/inspector/use-docx-tab-edit-session.ts": 2,
       "apps/web/src/components/inspector/use-file-tab-rename.ts": 1,
@@ -498,12 +492,11 @@
       "apps/web/src/components/search-dialog.tsx": 1,
       "apps/web/src/components/sidebar-user-menu.tsx": 1,
       "apps/web/src/components/translate-document-dialog.tsx": 1,
+      "apps/web/src/features/chat/hooks/use-chat-session.ts": 5,
+      "apps/web/src/features/chat/hooks/use-workspace-chat-mention-registration.ts": 2,
+      "apps/web/src/features/chat/queries.ts": 1,
       "apps/web/src/lib/prompts/use-saved-prompts.ts": 1,
-      "apps/web/src/router.tsx": 1,
       "apps/web/src/routes/_protected.chat/-components/chat-thread-page.tsx": 3,
-      "apps/web/src/routes/_protected.chat/-hooks/use-chat-session.ts": 5,
-      "apps/web/src/routes/_protected.chat/-hooks/use-workspace-chat-mention-registration.ts": 2,
-      "apps/web/src/routes/_protected.chat/-queries.ts": 1,
       "apps/web/src/routes/_protected.chat/index.tsx": 4,
       "apps/web/src/routes/_protected.contacts/$contactId.tsx": 2,
       "apps/web/src/routes/_protected.contacts/-components/contact-caches.ts": 1,
@@ -517,8 +510,6 @@
       "apps/web/src/routes/_protected.knowledge/-components/template-field-mention-extension.ts": 1,
       "apps/web/src/routes/_protected.knowledge/-components/template-form.tsx": 1,
       "apps/web/src/routes/_protected.knowledge/-components/template-prefill-panel.tsx": 1,
-      "apps/web/src/routes/_protected.knowledge/-components/template-studio-chat.tsx": 6,
-      "apps/web/src/routes/_protected.knowledge/-components/template-studio-suggestions.ts": 1,
       "apps/web/src/routes/_protected.knowledge/tools.tsx": 1,
       "apps/web/src/routes/_protected.settings/-components/account/sessions-card.tsx": 1,
       "apps/web/src/routes/_protected.settings/-components/organization/ai-config-card.tsx": 2,
@@ -532,7 +523,7 @@
       "apps/web/src/routes/_protected.settings/organization.route.tsx": 1,
       "apps/web/src/routes/_protected.settings/route.tsx": 1,
       "apps/web/src/routes/_protected.todos/index.tsx": 4,
-      "apps/web/src/routes/_protected.tsx": 4,
+      "apps/web/src/routes/_protected.tsx": 3,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/-components/billing/rate-management-dialog.tsx": 1,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/-components/flows/run-launcher.tsx": 1,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/-components/members-section.tsx": 1,
@@ -541,7 +532,6 @@
       "apps/web/src/routes/_protected.workspaces/$workspaceId/-components/view/view-toolbar.tsx": 1,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/-hooks/use-create-b-boxes.ts": 1,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/-hooks/use-start-workflow.ts": 1,
-      "apps/web/src/routes/_protected.workspaces/$workspaceId/route.tsx": 1,
       "apps/web/src/routes/_protected.workspaces/-components/create-matter-dialog.tsx": 3,
       "apps/web/src/routes/_protected.workspaces/-components/matter-context-menu.tsx": 1,
       "apps/web/src/routes/auth/organization.tsx": 1,
